+++ app/src/main/java/com/stratagile/pnrouter/ui/activity/main/EncryptMsgActivity.kt
@@ -22,15 +22,12 @@ import com.stratagile.pnrouter.ui.activity.main.component.DaggerEncryptMsgCompon
 import com.stratagile.pnrouter.ui.activity.main.contract.EncryptMsgContract
 import com.stratagile.pnrouter.ui.activity.main.module.EncryptMsgModule
 import com.stratagile.pnrouter.ui.activity.main.presenter.EncryptMsgPresenter
-import com.stratagile.pnrouter.utils.FileMangerDownloadUtils
-import com.stratagile.pnrouter.utils.LibsodiumUtil
-import com.stratagile.pnrouter.utils.RxEncodeTool
-import com.stratagile.pnrouter.utils.SpUtil
+import com.stratagile.pnrouter.utils.*
 import com.stratagile.pnrouter.view.SweetAlertDialog
 import com.stratagile.tox.toxcore.KotlinToxService
 import com.stratagile.tox.toxcore.ToxCoreJni
 import kotlinx.android.synthetic.main.emailname_bar.*
-import kotlinx.android.synthetic.main.encrypt_reg_activity.*
+import kotlinx.android.synthetic.main.encrypt_main_activity.*
 import kotlinx.android.synthetic.main.encrypt_source.*
 import org.libsodium.jni.Sodium
 
@@ -53,103 +50,16 @@ class EncryptMsgActivity : BaseActivity(), EncryptMsgContract.View {
     internal lateinit var mPresenter: EncryptMsgPresenter
     var verifiers= arrayListOf<String>()
     override fun onCreate(savedInstanceState: Bundle?) {
+        needFront = true
         super.onCreate(savedInstanceState)
     }
 
     override fun initView() {
-        setContentView(R.layout.encrypt_reg_activity)
+        setContentView(R.layout.encrypt_main_activity)
     }
     override fun initData() {
 
-        title.text = "Encrypt And Decrypt"
-        regeditBtn.setOnClickListener {
-            var emaiLAccount = account_editText.text.toString()
-            if(emaiLAccount =="")
-            {
-                toast(R.string.Account_cannot_be_empty)
-                return@setOnClickListener
-            }
-            val qlcClient = QlcClient(ConstantValue.qlcNode)
-            val rpc = DpkiRpc(qlcClient)
-            verifiers= arrayListOf<String>()
-            Thread(Runnable() {
-                run() {
-                    runOnUiThread {
-                        showProgressDialog(getString(R.string.waiting))
-                    }
-                    var getAllVerifiersResult =  rpc.getAllVerifiers(null)
-                    if(getAllVerifiersResult != null)
-                    {
-                        var firstResult = getAllVerifiersResult.get("result") as JSONArray
-                        for (i in 0..(firstResult.size - 1)){
-                            if(i>4)
-                            {
-                                break;
-                            }
-                            var obj:JSONObject = firstResult.get(i) as JSONObject
-                            var obj_account = obj.getString("account")
-                            var obj_type = obj.getString("type")
-                            var obj_id = obj.getString("id")
-                            verifiers.add(obj_account)
-                        }
-                        var qlcAccountEntityList = AppConfig.instance.mDaoMaster!!.newSession().qlcAccountDao.queryBuilder().where(QLCAccountDao.Properties.IsCurrent.eq(true)).list()
-                        if(qlcAccountEntityList != null && qlcAccountEntityList.size > 0)
-                        {
-                            var qlcAccount = qlcAccountEntityList.get(0)
-                            val publicBlock = JSONArray()
-                            var stateBlock = PublishBlock()
-                            stateBlock.account = qlcAccount.address  //"id" -> "test2.dpk@qlc.com"
-                            stateBlock.type ="email"
-                            stateBlock.id = emaiLAccount
-                            stateBlock.fee = "500000000"
-                            var pubkey = Helper.byteToHexString(RxEncodeTool.base64Decode(ConstantValue.libsodiumpublicMiKey)).toLowerCase();
-                            pubkey = pubkey.substring(0,64)
-                            stateBlock.pubkey = pubkey;
-                            stateBlock.verifiers = verifiers;
-                            publicBlock.add(JSONObject.parseObject(Gson().toJson(stateBlock)))
-                            val publicBlock2 = JSONArray()
-                            publicBlock2.add(publicBlock)
-                            publicBlock2.add(qlcAccount.privKey)
-                            try {
-                                var getPublishBlockResult =  rpc.getPublishBlockAndProcess(publicBlock2)
-                                var aa =""
-                            }catch (e:Exception)
-                            {
-                                e.printStackTrace()
-                                runOnUiThread {
-                                    closeProgressDialog()
-                                }
-                                return@Runnable
-                            }
-
-                            //var getPublishBlockResult =  rpc.getPublishBlock(publicBlock)
-                            /* var firstResult = getPublishBlockResult.get("result")  as JSONObject
-                             var verifiers = firstResult.get("verifiers")
-                             var block = firstResult.get("block")*/
-
-                            runOnUiThread {
-                                closeProgressDialog()
-                                showDialog("seed:",qlcAccount.seed)
-                            }
-                            /* val verifiersBlock = JSONArray()
-                             verifiersBlock.add("email")
-                             verifiersBlock.add(emaiLAccount)
-                             var getPubKeyByTypeAndIDResult =  rpc.getPubKeyByTypeAndID(verifiersBlock)
-
-
-                             val verifiersBlock22 = JSONArray()
-                             verifiersBlock22.add(qlcAccount.address)
-                             verifiersBlock22.add("email")
-                             var getPubKeyByTypeAndIDResult2 =  rpc.getPublishInfosByAccountAndType(verifiersBlock22)
-                             var bb = ""*/
-                        }
-
-                    }
-                }
-            }).start()
-
-
-        }
+        tvTitle.text = "Encrypt And Decrypt"
         decryptBtn.setOnClickListener {
             var password_editText = password_editText.text.toString()
             if(password_editText =="")
@@ -186,13 +96,21 @@ class EncryptMsgActivity : BaseActivity(), EncryptMsgContract.View {
                 var sourceTxt = LibsodiumUtil.DecryptFriendMsg2(miTxt,random_nonce,ConstantValue.libsodiumprivateMiKey!!,libsodiumpublicTemKey)
                 runOnUiThread {
                     closeProgressDialog()
-                    showDialog("Original text:",sourceTxt)
+                    if(sourceTxt !="")
+                    {
+                        showDialog("Original text:",sourceTxt)
+                    }else{
+                        toast(R.string.Decryption_failed)
+                    }
                 }
             }catch (e:Exception)
             {
                 toast(R.string.Decryption_failed)
                 return@setOnClickListener
             }
+        }
+        registerBtn.setOnClickListener {
+            startActivity(Intent(this, EncryptMsgTypeActivity::class.java))
         }
         encryptBtn.setOnClickListener {
             var emaiLAccount = account_editText.text.toString()
@@ -207,7 +125,10 @@ class EncryptMsgActivity : BaseActivity(), EncryptMsgContract.View {
                 toast(R.string.Please_enter_content)
                 return@setOnClickListener
             }
-
+            if (!NetUtils.isNetworkAvalible(this)) {
+                toast(getString(R.string.internet_unavailable))
+                return@setOnClickListener
+            }
             val qlcClient = QlcClient(ConstantValue.qlcNode)
             val rpc = DpkiRpc(qlcClient)
             Thread(Runnable() {
@@ -218,56 +139,65 @@ class EncryptMsgActivity : BaseActivity(), EncryptMsgContract.View {
                     runOnUiThread {
                         showProgressDialog(getString(R.string.waiting))
                     }
-                    var getPubKeyByTypeAndIDResult =  rpc.getPubKeyByTypeAndID(publicBlock)
-                    var pubKey = ""
-                    if(getPubKeyByTypeAndIDResult != null)
-                    {
-                        var firstResult = getPubKeyByTypeAndIDResult.get("result")  as JSONArray
-                        if(firstResult.size >0)
+                    try {
+                        var getPubKeyByTypeAndIDResult =  rpc.getPubKeyByTypeAndID(publicBlock)
+                        var pubKey = ""
+                        if(getPubKeyByTypeAndIDResult != null)
                         {
-                            for (i in 0..(firstResult.size - 1)) {
-                                var obj:JSONObject = firstResult.get(i) as JSONObject
-                                var obj_type = obj.getString("type")
-                                var obj_pubKey = obj.getString("pubKey")
-                                if(obj_type =="email")
-                                {
-                                    pubKey = obj_pubKey;
-                                    break
+                            var firstResult = getPubKeyByTypeAndIDResult.get("result")  as JSONArray
+                            if(firstResult.size >0)
+                            {
+                                for (i in 0..(firstResult.size - 1)) {
+                                    var obj:JSONObject = firstResult.get(i) as JSONObject
+                                    var obj_type = obj.getString("type")
+                                    var obj_pubKey = obj.getString("pubKey")
+                                    if(obj_type =="email")
+                                    {
+                                        pubKey = obj_pubKey;
+                                        break
+                                    }
+                                }
+                                var dst_public_TemKey_My = ByteArray(32)
+                                var dst_private_Temkey_My = ByteArray(32)
+                                var crypto_box_keypair_Temresult = Sodium.crypto_box_keypair(dst_public_TemKey_My,dst_private_Temkey_My)
+                                var libsodiumprivateTemKeyNew = RxEncodeTool.base64Encode2String(dst_private_Temkey_My)
+                                var libsodiumpublicTemKeyNew  =  RxEncodeTool.base64Encode2String(dst_public_TemKey_My)
+                                /* if(BuildConfig.DEBUG){
+                                     libsodiumprivateTemKeyNew = ConstantValue.libsodiumprivateTemKey!!
+                                     libsodiumpublicTemKeyNew = ConstantValue.libsodiumpublicTemKey!!
+                                 }*/
+                                var friendMiPublic = Helper.hexStringToBytes(pubKey)
+                                var msgMap = LibsodiumUtil.EncryptSendMsg(password_editText,friendMiPublic,ConstantValue.libsodiumprivateSignKey!!,libsodiumprivateTemKeyNew,libsodiumpublicTemKeyNew,ConstantValue.libsodiumpublicMiKey!!)
+                                var minTxt = msgMap.get("encryptedBase64") as String
+                                var NonceBase64 =  msgMap.get("NonceBase64") as String
+                                var msgSouce = LibsodiumUtil.DecryptMyMsg(minTxt, NonceBase64, msgMap.get("dst_shared_key_Mi_My64") as String, ConstantValue.libsodiumpublicMiKey!!, ConstantValue.libsodiumprivateMiKey!!)
+                                var miStrBegin ="UUxDSUQ="+libsodiumpublicTemKeyNew+"00"+NonceBase64+minTxt
+                                runOnUiThread {
+                                    closeProgressDialog()
+                                    showDialog("Ciphertext:",miStrBegin)
+                                }
+                            }else{
+                                runOnUiThread {
+                                    closeProgressDialog()
+                                    toast(R.string.Please_registe_your_account_first)
                                 }
                             }
-                            var dst_public_TemKey_My = ByteArray(32)
-                            var dst_private_Temkey_My = ByteArray(32)
-                            var crypto_box_keypair_Temresult = Sodium.crypto_box_keypair(dst_public_TemKey_My,dst_private_Temkey_My)
-                            var libsodiumprivateTemKeyNew = RxEncodeTool.base64Encode2String(dst_private_Temkey_My)
-                            var libsodiumpublicTemKeyNew  =  RxEncodeTool.base64Encode2String(dst_public_TemKey_My)
-                           /* if(BuildConfig.DEBUG){
-                                libsodiumprivateTemKeyNew = ConstantValue.libsodiumprivateTemKey!!
-                                libsodiumpublicTemKeyNew = ConstantValue.libsodiumpublicTemKey!!
-                            }*/
-                            var friendMiPublic = Helper.hexStringToBytes(pubKey)
-                            var msgMap = LibsodiumUtil.EncryptSendMsg(password_editText,friendMiPublic,ConstantValue.libsodiumprivateSignKey!!,libsodiumprivateTemKeyNew,libsodiumpublicTemKeyNew,ConstantValue.libsodiumpublicMiKey!!)
-                            var minTxt = msgMap.get("encryptedBase64") as String
-                            var NonceBase64 =  msgMap.get("NonceBase64") as String
-                            var msgSouce = LibsodiumUtil.DecryptMyMsg(minTxt, NonceBase64, msgMap.get("dst_shared_key_Mi_My64") as String, ConstantValue.libsodiumpublicMiKey!!, ConstantValue.libsodiumprivateMiKey!!)
-                            var miStrBegin ="UUxDSUQ="+libsodiumpublicTemKeyNew+"00"+NonceBase64+minTxt
-                            runOnUiThread {
-                                closeProgressDialog()
-                                showDialog("Ciphertext:",miStrBegin)
-                            }
+
                         }else{
                             runOnUiThread {
                                 closeProgressDialog()
                                 toast(R.string.Decryption_failed)
                             }
-                        }
 
-                    }else{
+                        }
+                    }catch (e:Exception)
+                    {
                         runOnUiThread {
                             closeProgressDialog()
                             toast(R.string.Decryption_failed)
                         }
-
                     }
+
                 }
             }).start()
         }
