+++ app/src/main/res/layout/encrypt_reg_activity.xml
@@ -12,14 +12,14 @@
     <ImageView
         android:id="@+id/emailLogo"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="@dimen/x250"
         android:layout_marginTop="@dimen/x50"
         android:src="@mipmap/tabbar_album_encryption_a" />
     <LinearLayout
         android:id="@+id/llCircle"
         android:layout_width="match_parent"
         android:layout_height="@dimen/x108"
-        android:layout_marginLeft="@dimen/x43"
+        android:layout_marginLeft="@dimen/x40"
         android:layout_marginRight="@dimen/x40"
         android:layout_marginTop="4dp"
         android:background="@drawable/btn_maincolor5"
@@ -65,6 +65,7 @@
     <TextView
         android:id="@+id/regeditBtn"
         android:layout_marginRight="@dimen/x50"
+        android:layout_marginTop="@dimen/x50"
         android:layout_marginLeft="@dimen/x50"
         android:gravity="center"
         android:layout_width="match_parent"
@@ -74,57 +75,6 @@
         android:textColor="@color/white"
         android:textSize="@dimen/commemTextSize"
         />
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Encryption and decryption content："
-        android:textColor="@color/emailtextColor"
-        android:layout_marginTop="@dimen/x40"
-        android:layout_marginLeft="@dimen/x40"
-        android:textSize="16sp"
-        />
-    <include layout="@layout/encrypt_source" />
-
-
-
-
-
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <TextView
-            android:id="@+id/encryptBtn"
-            android:layout_weight="1"
-            android:layout_marginRight="@dimen/x10"
-            android:layout_marginLeft="@dimen/x10"
-            android:gravity="center"
-            android:layout_width="wrap_content"
-            android:layout_height="@dimen/x58"
-            android:background="@drawable/btn_purcle"
-            android:text="@string/Encrypt"
-            android:textColor="@color/white"
-            android:textSize="@dimen/commemTextSize"
-            />
-        <TextView
-            android:id="@+id/decryptBtn"
-            android:layout_marginRight="@dimen/x10"
-            android:layout_marginLeft="@dimen/x10"
-            android:layout_weight="1"
-            android:gravity="center"
-            android:layout_width="wrap_content"
-            android:layout_height="@dimen/x58"
-            android:background="@drawable/btn_maincolor"
-            android:text="@string/Decrypt"
-            android:textColor="@color/white"
-            android:textSize="@dimen/commemTextSize"
-            />
-
-    </LinearLayout>
-
-
 
 </LinearLayout>
 </android.support.v4.widget.NestedScrollView>
